lam/requirements.txt
# Core dependencies
requests>=2.31.0
cryptography>=41.0.0

# Phase 3: Production & Advanced Features
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# Voice interface
SpeechRecognition>=3.10.0
pyttsx3>=2.90

# Phase 4: Enterprise Features
# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.12.0

# GraphQL
strawberry-graphql[fastapi]>=0.214.0

# Machine Learning
numpy>=1.24.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
